{"schema":"decodex/commit/1","summary":"Split structured retrieval modules","authority":"manual"}
